// src/types/weather.types.ts
export interface WeatherData {
  name: string;
  main: {
    temp: number;
    feels_like: number;
    humidity: number;
  };
  weather: {
    description: string;
    icon: string;
  }[];
  wind: {
    speed: number;
  };
  sys: {
    country: string;
  };
}

export interface WeatherState {
  data: WeatherData | null;
  loading: boolean;
  error: string | null;
}